// frontend/src/interface/common/enums/enums.ts
// src/interface/common/enums.ts

export enum Sex {
    Male = "MASCULINO",
    Female = "FEMENINO",
    Other = "OTROS",
  }
  
  export enum Source {
    SocialMedia = "REDES SOCIALES",
    InPerson = "CAPTADOR",
  }
  
  export enum CommunicationPreference {
    Whatsapp = "WHATSAPP",
    Phone = "TELÉFONO",
  }
  
  export enum School {
    Quinta = "QUINTA NORMAL",
    Buin = "BUÍN",
    Granja = "LA GRANJA",
    Nunoa = "ÑUÑOA",
    Pudahuel = "PUDAHUEL",
    Miguel = "SAN MIGUEL",
  }
  
  export enum Course {
    NB1 = "1° NIVEL BÁSICO",
    NB2 = "2° NIVEL BÁSICO",
    NB3 = "3° NIVEL BÁSICO",
    NM1 = "1° NIVEL MEDIO",
    NM2 = "2° NIVEL MEDIO",
  }